+++ dao/spirit_dao_mongo_test.go
@@ -1,0 +1,79 @@
+package dao
+
+import (
+	"github.com/sebastienfr/handsongo/model"
+	"testing"
+	"time"
+)
+
+const (
+	mongoCnx = "mongodb://192.168.99.100/spirits"
+)
+
+func TestDAOMongo(t *testing.T) {
+	daoMongo, err := GetSpiritDAO(mongoCnx, DAOMongo)
+	if err != nil {
+		t.Error(err)
+	}
+
+	toSave := model.Spirit{
+		Name:         "Caroni",
+		Distiller:    "Caroni",
+		Bottler:      "Velier",
+		Country:      "Trinidad",
+		Composition:  "Melasse",
+		SpiritType:   model.TypeRhum,
+		Age:          15,
+		BottlingDate: time.Date(2015, 01, 01, 0, 0, 0, 0, time.UTC),
+		Score:        8.5,
+		Comment:      "heavy tire taste",
+	}
+
+	err = daoMongo.SaveSpirit(&toSave)
+	if err != nil {
+		t.Error(err)
+	}
+
+	t.Log("initial spirit saved", toSave)
+
+	spirits, err := daoMongo.GetAllSpirits(NoPaging, NoPaging)
+	if err != nil {
+		t.Error(err)
+	}
+
+	t.Log("initial spirit found all", spirits[0])
+
+	oneSpirit, err := daoMongo.GetSpiritByID(spirits[0].ID.Hex())
+	if err != nil {
+		t.Error(err)
+	}
+
+	t.Log("initial spirit found one", oneSpirit)
+
+	oneSpirit.Age = 18
+	oneSpirit.Comment = "soft tarmac smell"
+	chg, err := daoMongo.UpsertSpirit(oneSpirit.ID.Hex(), oneSpirit)
+	if err != nil {
+		t.Error(err)
+	}
+
+	t.Log("initial spirit modified", chg, oneSpirit)
+
+	oneSpirit, err = daoMongo.GetSpiritByID(oneSpirit.ID.Hex())
+	if err != nil {
+		t.Error(err)
+	}
+
+	t.Log("initial spirit found one modified", oneSpirit)
+
+	err = daoMongo.DeleteSpirit(oneSpirit.ID.Hex())
+	if err != nil {
+		t.Error(err)
+	}
+
+	oneSpirit, err = daoMongo.GetSpiritByID(oneSpirit.ID.Hex())
+	if err != nil {
+		t.Log("initial spirit deleted", err, oneSpirit)
+	}
+
+}