+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ## v0.0.1 [Unreleased]
-- 16/05/10 feat(all): project start (SFR)
+- 16/05/11 feat(dao): add model of spirits, mongo dao, factory and update vendors (SFR)
+- 16/05/10 feat(all): project start, add command line, vendors, makefile, docker, logger, tests and statistics (SFR)